requirements.txt
python-telegram-bot
python-dotenv
pandas
openpyxl
APScheduler